+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [4.18.2](https://github.com/customerio/customerio-android/compare/4.18.1...4.18.2) (2026-06-22)
+
+### Bug Fixes
+
+* scope in-app polling to process lifecycle instead of per-activity ([#746](https://github.com/customerio/customerio-android/issues/746)) ([5d30ba1](https://github.com/customerio/customerio-android/commit/5d30ba19f412884ca109539486d14bae9b843e9e))
+
+## [4.18.1](https://github.com/customerio/customerio-android/compare/4.18.0...4.18.1) (2026-06-15)
+
+### Bug Fixes
+
+* harden push notification activity and in-app message WebView ([#733](https://github.com/customerio/customerio-android/issues/733)) ([1323529](https://github.com/customerio/customerio-android/commit/13235298de8a3eb0c9ece2fe2c9eab677cd65689))
+
 ## [4.18.0](https://github.com/customerio/customerio-android/compare/4.17.1...4.18.0) (2026-06-04)
 
 ### Features